engineer/pyproject.toml
[project]
name = "engineer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "tavily-python>=0.7.17",
    "langchain>=1.2.0",
    "lanchain-openai>=1.1.7",
    "pydantic>=2.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]
